epleads/requirements.txt
asgiref==3.4.1
attrs==21.4.0
Automat==20.2.0
cffi==1.15.0
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.1
django-compat==1.0.15
django-import-export==2.7.1
django-phonenumber-field==6.0.0
et-xmlfile==1.1.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.7.1
MarkupPy==1.14
mysqlclient==2.1.0
odfpy==1.4.1
openpyxl==3.0.9
parsel==1.6.0
phonenumberslite==8.12.42
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==21.0.0
PyYAML==6.0
queuelib==1.6.2
Scrapy==2.5.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typing-extensions==4.0.1
tzdata==2021.5
w3lib==1.22.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
